warning
svn:sync-xref-src-repo-rev: r85556

+++ src/VBox/Main/src-client/EbmlWriter.cpp
@@ -1,4 +1,4 @@
-/* $Id: EbmlWriter.cpp 45922 2013-05-06 19:11:54Z noreply@oracle.com $ */
+/* $Id: EbmlWriter.cpp 45924 2013-05-06 19:34:30Z noreply@oracle.com $ */
 /** @file
  * EbmlWriter.cpp - EBML writer + WebM container
  */
@@ -417,7 +417,7 @@ int Ebml_WriteWebMBlock(EbmlGlobal                *glob,
     if (RT_SUCCESS(rc))
         rc = ebml_WriteID(glob, SimpleBlock);
 
-    uint64_t block_length = pkt->data.frame.sz + 4;
+    uint32_t block_length = pkt->data.frame.sz + 4;
     block_length |= 0x10000000;
     if (RT_SUCCESS(rc))
         rc = ebml_WriteU32(glob, RT_H2BE_U32(block_length));
